langchain/Chain_Prompt_Memory_Retriever/requirements.txt
langchain
langchain-openai
python-dotenv
chromadb
pypdf
langchain-community